refactor(constants): move browser-only nav buttons out of shared
