+++ src/app/(preview)/preview/[id]/_components/Navigation.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { Button } from "@/client/components/atoms";
-import { navigationButtons } from "@/shared/constants";
+import { navigationButtons } from "@/client/constants/navigation";
 import type { NavigationGeo } from "@/client/hooks";
 import Image from "next/image";
 import React from "react";
